# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clínica Serena API
  description: |
    Contrato oficial de la API de Clínica Serena. Es la fuente de verdad
    compartida entre el backend (Spring Boot), el frontend web (Next.js) y
    la app Android del paciente.

    Define los endpoints iniciales, agrupados en Health, Catálogo, Auth,
    Citas, Portal del paciente y Personal (ver tags).

    ## Estado de implementación
    - **Implementado en esta fase**: entidades y migración de paciente, cuenta
      y sesión; `POST /auth/login`, `GET /auth/me` y `POST /auth/logout` con
      BCrypt y Bearer opaco revocable.
    - **Implementado en esta fase**: autenticación Bearer separada para
      personal, creación administrativa de cuentas de recepción/médico,
      agenda persistida y registro de llegada con auditoría.
    - **Implementado para acceso web unificado**: `POST /auth/login-unified`
      resuelve la cuenta por correo y contraseña sin aceptar un rol del cliente.
      Si el mismo correo existe en paciente y personal, rechaza el acceso de
      forma genérica hasta que esa ambigüedad sea resuelta administrativamente.
    - **Implementado en esta fase**: vinculación administrativa de cuentas
      MEDICO con profesionales del catálogo, citas propias del profesional,
      expediente básico, historial de atenciones y registro inmutable de
      atención con diagnóstico y receta (`/medico/**`).
    - **Implementado para el portal del paciente**: perfil propio consultable y
...

...
servers:
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Inicia sesión como paciente
      description: |
        Endpoint público. Devuelve un token de acceso que debe enviarse como
        `Authorization: Bearer <token>` en los endpoints protegidos.
        La contraseña se verifica con BCrypt. La respuesta contiene un token
        Bearer opaco de vida limitada; el servidor persiste únicamente su hash
        y puede revocarlo con `POST /auth/logout`. El límite local predeterminado
        permite cinco fallos por correo y dirección remota en 60 segundos y
        aplica una pausa de 60 segundos antes de aceptar nuevos intentos.
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Datos de entrada inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

        "401":
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Consulta la identidad del paciente autenticado
      description: Requiere un Bearer opaco vigente emitido por el login y el rol PACIENTE.
      security:
        - bearerAuth: []
      operationId: getCurrentPatientIdentity
      responses:
        "200":
          description: Identidad autenticada, sin datos clínicos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientIdentity"
        "401":
          description: Token ausente, inválido, expirado o revocado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoca la sesión Bearer actual
      description: Requiere el rol PACIENTE. La revocación es inmediata.
      security:
        - bearerAuth: []
      operationId: logout
      responses:
        "204":
          description: Sesión revocada.
        "401":
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /staff/auth/login:
    post:
      tags: [Personal]
      summary: Inicia sesión como personal
      description: Login público para cuentas ADMIN, RECEPCION o MEDICO. Devuelve un Bearer opaco separado del de pacientes.
      security: []
      operationId: loginStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginRequest"
      responses:
        "200":
          description: Login exitoso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /staff/auth/me:
    get:
      tags: [Personal]
      summary: Consulta la identidad del personal autenticado
      security:
        - bearerAuth: []
      operationId: getCurrentStaffIdentity
      responses:
        "200":
          description: Identidad y rol autenticados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffIdentity"
        "401":
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /staff/auth/logout:
    post:
      tags: [Personal]
      summary: Revoca la sesión del personal
      security:
        - bearerAuth: []
      operationId: logoutStaff
      responses:
        "204": { description: Sesión revocada. }
        "401":
          description: No autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /staff/accounts:
    post:
      tags: [Personal]
      summary: Crea una cuenta habilitada de recepción o médico
      description: Requiere rol ADMIN. No permite crear el primer ADMIN desde el API.
      security:
        - bearerAuth: []
      operationId: createStaffAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffAccountRequest"
      responses:
        "201":
          description: Cuenta habilitada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffAccount"
        "403": { description: Solo ADMIN puede crear cuentas de personal. }
        "409": { description: El correo ya existe. }
    get:
      tags: [Personal]
      summary: Lista las cuentas de personal con su estado de vinculación
      description: Requiere rol ADMIN. Las cuentas MEDICO sin profesional aparecen como `PENDIENTE_VINCULACION`.
      security:
        - bearerAuth: []
      operationId: listStaffAccounts
      parameters:
        - name: role
          in: query
          required: false
          schema: { $ref: "#/components/schemas/StaffRole" }
      responses:
        "200":
          description: Cuentas ordenadas por nombre.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StaffAccount" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /staff/accounts/{accountId}/practitioner:
    parameters:
      - name: accountId
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Personal]
      summary: Asigna o corrige el profesional de una cuenta MEDICO
      description: |
        Requiere rol ADMIN; el administrador que actúa se toma del principal.
        La operación bloquea la cuenta y el profesional. La base de datos impide
        que un profesional quede asignado a dos cuentas activas incluso ante
        solicitudes concurrentes. Repetir la misma asignación no genera historial;
        cada cambio se registra en `historial_vinculacion_medico`.
      security:
        - bearerAuth: []
      operationId: linkStaffAccountPractitioner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkPractitionerRequest" }
      responses:
        "200":
          description: Cuenta vinculada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffAccount" }
        "400":
          description: Cuerpo inválido o la cuenta no es MEDICO (`STAFF_ACCOUNT_NOT_MEDICAL`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: La cuenta (`STAFF_ACCOUNT_NOT_FOUND`) o el profesional (`PRACTITIONER_NOT_FOUND`) no existe.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: El profesional ya está asignado a otra cuenta activa (`PRACTITIONER_ALREADY_LINKED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    delete:
      tags: [Personal]
      summary: Retira la vinculación de una cuenta MEDICO
      description: Requiere rol ADMIN. Conserva las atenciones previas y registra el cambio en el historial.
      security:
        - bearerAuth: []
      operationId: unlinkStaffAccountPractitioner
      responses:
        "200":
          description: Cuenta sin vinculación.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffAccount" }
        "400":
          description: La cuenta no es MEDICO (`STAFF_ACCOUNT_NOT_MEDICAL`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: La cuenta no existe (`STAFF_ACCOUNT_NOT_FOUND`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: La cuenta no tenía profesional asignado (`PRACTITIONER_NOT_LINKED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /staff/agenda:
    get:
      tags: [Personal]
      summary: Consulta la agenda persistida
      description: Requiere rol ADMIN o RECEPCION. El resultado está limitado a 100 filas.
      security:
        - bearerAuth: []
      operationId: listStaffAgenda
      parameters:
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AppointmentStatus" }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Citas persistidas ordenadas por fecha.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ReceptionAppointment" }
        "403": { description: Rol sin permiso para agenda. }
  /staff/agenda/{appointmentId}/arrival:
    post:
      tags: [Personal]
      summary: Registra la llegada de un paciente
      description: Requiere RECEPCION. La cita se bloquea durante la operación y una segunda llegada devuelve conflicto.
      security:
        - bearerAuth: []
      operationId: registerAppointmentArrival
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Llegada persistida con actor y hora.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceptionAppointment" }
        "404": { description: La cita no existe. }
        "409": { description: La llegada ya existe o la cita no admite llegada. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Token opaco emitido por `POST /auth/login`. Solo se almacena su hash.
        Expira en 1800 segundos por defecto y se revoca con `POST /auth/logout`.
...
  responses:
    Unauthenticated:
      description: Sin sesión válida (`UNAUTHENTICATED`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: La sesión es válida pero su rol no tiene permiso (`FORBIDDEN`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    StaffRole:
      type: string
      enum: [ADMIN, RECEPCION, MEDICO]
    StaffLoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 12, maxLength: 72 }
    StaffIdentity:
      type: object
      required: [accountId, email, fullName, role, practitionerLinkStatus]
      properties:
        accountId: { type: string }
        email: { type: string, format: email }
        fullName: { type: string }
        role: { $ref: "#/components/schemas/StaffRole" }
        practitionerLinkStatus: { $ref: "#/components/schemas/PractitionerLinkStatus" }
        practitionerId: { type: string, nullable: true }
    PractitionerLinkStatus:
      type: string
      enum: [VINCULADA, PENDIENTE_VINCULACION, NO_APLICA]
      description: |
        Derivado de la relación persistida. `PENDIENTE_VINCULACION` identifica
        cuentas MEDICO que aún no pueden acceder a datos clínicos; `NO_APLICA`
        corresponde a ADMIN y RECEPCION.
    LinkPractitionerRequest:
      type: object
      additionalProperties: false
      required: [practitionerId]
      properties:
        practitionerId: { type: string, maxLength: 36 }
    CreateStaffAccountRequest:
      type: object
      additionalProperties: false
      required: [email, fullName, role, password]
      properties:
        email: { type: string, format: email }
        fullName: { type: string, maxLength: 160 }
        role: { type: string, enum: [RECEPCION, MEDICO] }
        password: { type: string, format: password, minLength: 12, maxLength: 72 }
    StaffAccount:
      type: object
      required: [accountId, email, fullName, role, status, practitionerLinkStatus]
      properties:
        accountId: { type: string }
        email: { type: string, format: email }
        fullName: { type: string }
        role: { $ref: "#/components/schemas/StaffRole" }
        status: { type: string, enum: [ACTIVA, BLOQUEADA] }
        practitionerLinkStatus: { $ref: "#/components/schemas/PractitionerLinkStatus" }
        practitionerId: { type: string, nullable: true }
        practitionerName: { type: string, nullable: true }
        practitionerLinkedAt: { type: string, format: date-time, nullable: true }
...
    ReceptionAppointment:
      type: object
      required: [id, patientId, practitionerId, specialtyId, scheduledAt, status]
      properties:
        id: { type: string }
        patientId: { type: string }
        practitionerId: { type: string }
        specialtyId: { type: string }
        scheduledAt: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/AppointmentStatus" }
        arrivalAt: { type: string, format: date-time, nullable: true }
        arrivalByAccountId: { type: string, nullable: true }
...
    AppointmentStatus:
      type: string
      enum: [PENDIENTE, CONFIRMADA, CANCELADA, COMPLETADA]
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [accessToken, tokenType]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
...
    PatientIdentity:
      type: object
      required: [patientId, email, accountStatus]
      properties:
        patientId:
          type: string
          description: Identificador estable del paciente autenticado.
        email:
          type: string
          format: email
        accountStatus:
          type: string
          enum: [ACTIVA]
...
    ApiError:
      type: object
      required: [status, code, message]
      properties:
        status:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Datos inválidos
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: fecha
              message:
                type: string
                example: La fecha es obligatoria